pyproject.toml
[project]
name = "genai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt[standard]>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.12.0",
    "docx2txt>=0.9",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.135.1",
    "nomic>=3.9.0",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "pypdfium2>=5.6.0",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "tavily>=1.1.0",
    "uvicorn[standard]>=0.41.0",
]
